+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.1
+- Fixed ReadDiscount and CourseVania
+- Added Russian Language filter
+
 ## v1.0
 
 - Fresh start